+++ Unity SDK/Assets/Voximplant/UninitializedException.cs
@@ -8,6 +8,8 @@ using System.Runtime.Serialization;
 namespace Voximplant.Unity
 {
     /// <summary>
+    /// The exception that is thrown when the Voximplant Unity SDK is not initialized.
+    /// 
     /// <see cref="VoximplantSdk.Initialize(Voximplant.Unity.Client.ClientConfig)"/> must be called before accessing
     /// <see cref="VoximplantSdk.GetClient()"/>, <see cref="VoximplantSdk.GetAudioManager()"/> and <see cref="VoximplantSdk.GetCameraManager()"/>.
     /// </summary>